JobGenerator.java: Migrate from "master" node label to "built-in"
According to the changelog, Jenkins CI applies the "built-in" label for
the controller node instead of "master", how it was called before:
https://www.jenkins.io/changelog-stable/#v2.319.1

The plugin explicitly defines the "master" as the required label in
code. This behaviour prevents it to be used in new installation with
greater or equal to this Jenkins CI version without explicitly and
manually assigning this label to the controller node too.